feat(rust/llvm): Accept numbered types, layout quirks and the surviving const exprs
Fourth conformance pass, all three ratchets up: Assembler 175 to 180,
Verifier 121 to 125, differential 105 of 163 to 111 of 181.

On the parse side: a struct named by number (`%0 = type`), which needs
its own flag because a struct whose name is the word "0" prints quoted
and a numbered one prints bare; `align(4)` as well as `align 4"; the
deprecated sized-aggregate alignment in a datalayout; a phi with no
edges; inline string attributes on a global; more calling conventions;
an address space written before a call's type; and metadata integers
wider than 64 bits.

Measuring beat guessing on constant expressions: LLVM 21 removed most of
them but kept exactly add, sub and xor, so the note claiming all the
arithmetic ones were gone was wrong. They parse again.

Fixing the layout parser exposed eleven bitcast files whose real rule is
that a bitcast may change neither address space nor size. That now
applies to constant expressions as well as instructions.

+++ crates/llvm-ir-parse/src/metadata.rs
@@ -97,13 +97,13 @@ impl Parser {
             Token::Integer { negative, digits } => {
                 self.advance();
                 if negative {
-                    let value: i64 = digits.parse().map_err(|_| {
+                    let value: i128 = digits.parse().map_err(|_| {
                         self.error::<()>(format!("{digits} does not fit"))
                             .unwrap_err()
                     })?;
                     Ok(MdField::Signed(-value))
                 } else {
-                    let value: u64 = digits.parse().map_err(|_| {
+                    let value: u128 = digits.parse().map_err(|_| {
                         self.error::<()>(format!("{digits} does not fit"))
                             .unwrap_err()
                     })?;
